Abstract the signature system a bit.
The signature of a unit is the sum of powers of a base number, with each power representing a different fundamental unit.
Previously, the base was hardcoded to 20. Now it's a constant Qty.SIGNATURE_POWER = 32.

The list of fundamental dimensions is exposed as Qty.SIGNATURE_VECTOR.

The dictionary of known kinds of units was hardcoded as a list of signatures using base 20.
It's now calculated on startup using the new function unitSignatureFromDict.